My last styling commit
This commit makes minor changes to a bunch of styling across multiple pages.
The primary styling changes are to the Job Application Show page.
This page was really well done already, especially the reactivity and
the layout of the jsx. However, some things didn't match the mock ups
and some of the styling was a bit redundant. Tailwind sorts through
everything and removes redundancies when it compiles though, so that's
not an issue outside of maintaining code readability.

[Ticket: 70]